Move db_ctrl to a separate file
